feat: extract the map section and create the component `ResultMapSection`
