+++ src/app/search/utils/search.client.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { Activity, SearchResult } from '@action/activity';
+import { SearchResult } from '@action/activity';
 import {
   SearchPagination,
   SearchPaginationProps,
@@ -12,11 +12,11 @@ import {
   updateQueryString,
 } from '@components/search/SearchBar/fields/utils';
 import SearchProvider from '@components/search/SearchBar/SearchProvider';
-import SearchMapSection from '@components/search/SearchMapSection';
 import { zodResolver } from '@hookform/resolvers/zod';
 import useMediaQuery from '@hooks/use-media-query';
 import { useRouter, useSearchParams } from 'next/navigation';
-import { useMemo, useState } from 'react';
+import { useState } from 'react';
+import ResultMapSection from './ResultMapSection';
 import ResultSection from './ResultSection';
 
 type SearchClientProps = {
@@ -54,44 +54,6 @@ const SearchClient = ({ result }: SearchClientProps) => {
     });
     router.push(`/search?${updatedQuery}`);
   };
-
-  const mapMarkers = useMemo(() => {
-    const now = Date.now();
-
-    return activities
-      .filter(
-        (currentResult): currentResult is Activity =>
-          typeof currentResult.lat === 'number' &&
-          typeof currentResult.lng === 'number' &&
-          Array.isArray(currentResult.ticketPrice)
-      )
-      .map((activity) => {
-        const nearestTicket =
-          activity.ticketPrice.length > 0
-            ? activity.ticketPrice.reduce<{
-                name: string;
-                price: number;
-                startDateTime: string;
-                endDateTime: string;
-              }>((nearest, current) => {
-                const currentDiff = Math.abs(
-                  new Date(current.startDateTime).getTime() - now
-                );
-                const nearestDiff = Math.abs(
-                  new Date(nearest.startDateTime).getTime() - now
-                );
-                return currentDiff < nearestDiff ? current : nearest;
-              }, activity.ticketPrice[0])
-            : null;
-
-        return {
-          lat: activity.lat,
-          lng: activity.lng,
-          id: activity._id,
-          pricing: nearestTicket ? nearestTicket.price : 0,
-        };
-      });
-  }, [activities]);
 
   const toggleShow = () => {
     if (isMobile) {
@@ -137,6 +99,7 @@ const SearchClient = ({ result }: SearchClientProps) => {
               setCenterId={setCenterId}
               currentShow={currentShow}
             />
+
             <SearchPagination
               initialPage={currentPage}
               totalPage={totalPage}
@@ -147,16 +110,12 @@ const SearchClient = ({ result }: SearchClientProps) => {
           </div>
         ) : null}
 
-        {isMobile && currentShow === 'map' && (
-          <div className="h-fit w-full">
-            <SearchMapSection centerId={centerId} markers={mapMarkers} />
-          </div>
-        )}
-        {!isMobile && (
-          <div className="sticky top-0 h-fit w-full max-w-[26rem]">
-            <SearchMapSection centerId={centerId} markers={mapMarkers} />
-          </div>
-        )}
+        <ResultMapSection
+          activities={activities}
+          centerId={centerId}
+          isMobile={isMobile}
+          currentShow={currentShow}
+        />
       </section>
     </SearchProvider>
   );
